Add inputParser test
inputParser에 대한 클래스 추가
